example/requirements.txt
fastapi==0.115.1
tortoise-orm==0.24.2
aerich==0.7.2
sqlite-utils==3.36
pydantic==2.7.1
uvicorn==0.30.0
fastapi18n==0.1.0
google-cloud-translate==3.20.2
python-dotenv==1.0.1
Babel==2.16.0
contextvars
